pyproject.toml
[project]
name = "trgs_slam"
version = "0.1.0"
description = "A package for 3DGS based thermal SLAM."
authors = [{ name = "Spencer Carmichael", email = "specarmi@umich.edu" }]
requires-python = ">=3.9"

dependencies = [
    "torch>=2.7",
    "numpy>=2.2.6",
    "pypose>=0.7.5",
    "lietorch>=0.3",
    "lie_spline_torch>=0.1",
    "pyyaml>=6.0.3",
    "opencv-python-headless>=4.13.0.92",
    "tyro>=1.0.6",
    "gsplat>=1.5.1",
    "fused-ssim",
    "h5py>=3.15.1",
    "scipy>=1.15.3",
    "scikit-learn>=1.7.2",
    "torchmetrics[image]>=1.9.0",
    "evo>=1.34.3",
    "viser>=1.0.24",
    "nerfview>=0.1.2",
    "splines>=0.3.3"
]
